aimlchatbot/test_api.py
import requests
import json

url = "http://localhost:5001/api/chat"
payload = {"message": "Hello, how are you?"}
headers = {'Content-Type': 'application/json'}

try:
    response = requests.post(url, headers=headers, data=json.dumps(payload))
    print(f"Status Code: {response.status_code}")
    print(f"Response Body: {response.text}")
except Exception as e:
    print(f"Request failed: {e}")
