aimlchatbot/requirements.txt
fastapi
uvicorn
python-dotenv
huggingface-hub
httpx
pydantic
